LayoutTests/http/tests/iframe-monitor/eligibility-expected.txt

CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 4194304 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
Test resource monitor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector('iframe[name=frame1]').srcdoc is ""
PASS document.querySelector('iframe[name=frame2]').srcdoc is ""
PASS document.querySelector('iframe[name=frame3]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame4]').srcdoc is ""
PASS successfullyParsed is true

TEST COMPLETE
